refactor: add more describes on tests
