+++ tests/presentation/mocks/index.ts
@@ -1,3 +1,4 @@
 export * from './email-validator'
 export * from './mock-employee'
 export * from './mock-company'
+export * from './mock-signup'
