+++ tests/presentation/mocks/mock-signup.ts
@@ -1,0 +1,10 @@
+import { type SignUpController } from '../../../src/presentation/controllers/login/signup/signup'
+export const mockSignUpControllerRequest = (): SignUpController.Request => ({
+  email: 'any_mail@mail.com',
+  name: 'any_name',
+  password: 'any_password',
+  passwordConfirmation: 'any_password',
+  companyName: 'any_name',
+  cnpj: 'any_value',
+  cel_phone: 'any_cel_phone'
+})
